Фикс определения мп по сетям wifi Добавил запись в логМП информацию о wifi с координатами Увеличил кол-во мест откуда они вызываются

+++ app/src/main/java/ua/com/merchik/merchik/Options/Controls/OptionControlMP.java
@@ -211,7 +211,7 @@ public class OptionControlMP<T> extends OptionControl {
                                 stringBuilder.append("Поточне розташування присторою станом на ")
                                         .append(Clock.getHumanTime2(latestVpiLog.vpi))
                                         .append(" визначено на вiдстанi ")
-                                        .append(distance)
+                                        .append((int) distance)
                                         .append(" метрiв вiд ТТ");
                             }
                             click.onSuccess(stringBuilder.toString());
